geocode/requirements.txt
geopy==2.4.1
google-auth==2.48.0
graffiti-lookup-nyc==1.0.3
gspread==6.2.1
requests==2.28.0
